Refactor SortButton component: move to protectedAudience directory and update import path
